% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{break_response_ties}
\alias{break_response_ties}
\title{Break response ties}
\usage{
break_response_ties(data)
}
\arguments{
\item{data}{data.frame. Has two columns:
\itemize{
\item dose: numeric, contains untransformed doses
\item response: numeric, contains responses as inhibition
}}
}
\value{
input with responses broken, if any
}
\description{
If there are any ties, breaks them by adding an escalating 0.01 to EVERY
response
}
\examples{
no_ties <- data.frame(dose = 1:3, response = 1:3)
break_response_ties(no_ties)

ties <- data.frame(dose = 1:3, response = c(1, 1, 3))
break_response_ties(ties)
}
