% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bliss.R
\name{calculate_bliss}
\alias{calculate_bliss}
\title{Create Bliss independence matrix from dose-response matrix}
\usage{
calculate_bliss(mat)
}
\arguments{
\item{mat}{A dose response matrix, where values are \% inhibition}
}
\value{
A Bliss independence matrix, where values are \% inhibition
}
\description{
Create Bliss independence matrix from dose-response matrix
}
\details{
Values are limited between 0 and 100 before calculations are made
}
\examples{
rnorm(25, mean = 50, sd = 50, 5, 5) |> calculate_bliss()
}
