% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{clean_parameters}
\alias{clean_parameters}
\title{Detect and nudge suspicious looking parameters}
\usage{
clean_parameters(p, data)
}
\arguments{
\item{p}{A named vector of coefficients of a fit, with names "SLOPE", "MIN",
"MAX", and "IC50"}

\item{data}{data.frame. Has two columns:
\itemize{
\item dose: numeric, contains untransformed doses
\item response: numeric, contains responses as inhibition
}}
}
\description{
Detect and nudge suspicious looking parameters
}
