% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{convert_to_matrix}
\alias{convert_to_matrix}
\title{Convert dose-response data.frame into a matrix}
\usage{
convert_to_matrix(df)
}
\arguments{
\item{df}{data.frame with Conc1, Conc2, and Response}
}
\description{
Convert dose-response data.frame into a matrix
}
