% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_responses.R
\docType{data}
\name{dose_responses}
\alias{dose_responses}
\title{Dose response data for 3 different drug combinations}
\format{
An object of class \code{data.frame} with 66 rows and 7 columns.
}
\usage{
dose_responses
}
\description{
Dose response data for 3 different drug combinations
}
\keyword{datasets}
