% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Try to fit data to get initial parameters}
\usage{
estimate_parameters(x)
}
\value{
a object of class \code{drc}
}
\description{
Will attempt with log-logistic drc::LL.4. If fail or warn, falls back to
drc::L.4.
}
