% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnmf.R
\name{fit_cnmf}
\alias{fit_cnmf}
\title{Impute outliers using consensus non-negative matrix factorization.}
\usage{
fit_cnmf(data, has_outliers)
}
\arguments{
\item{data}{data.frame that has, minimally:
\itemize{
\item Response: numeric column that contains responses as viability
\item Conc1, Conc2: Columns containing concentrations of drug 1 and 2
}}

\item{has_outliers}{boolean which denotes if the data has outliers, as
determined by detect_outliers.}
}
\value{
numeric vector of imputed responses
}
\description{
Impute outliers using consensus non-negative matrix factorization.
}
\details{
Consensus is taken using venter mode of 120 rounds of estimation
}
