% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_dose_response}
\alias{fit_dose_response}
\title{Fit dose response data}
\usage{
fit_dose_response(data)
}
\arguments{
\item{data}{data.frame. Has two columns:
\itemize{
\item dose: numeric, contains untransformed doses
\item response: numeric, contains responses as inhibition
}}
}
\description{
Fit dose response data
}
\details{
Broadly, this function:
\enumerate{
\item Converts dose into log10 dose
\item Attempts a few preliminary fits to get good starting parameters
\item Checks if parameters seem unrealistic and nudges them accordingly
\item Does a final fit with these parameters
}
}
