% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_0_drug_highest}
\alias{make_0_drug_highest}
\title{Ensure 0 drug condition has highest viability}
\usage{
make_0_drug_highest(mat)
}
\arguments{
\item{mat}{a matrix of viabilities, where the top left entry is the 0, 0 drug
condition}
}
\description{
If the 0 drug condition is lower than either of its lowest-concentration
single agent drugs, set its viability to 100
}
