% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_outlier.R
\name{mark_outliers}
\alias{mark_outliers}
\title{Mark TRUE if values are outside of range}
\usage{
mark_outliers(values, iqrs)
}
\arguments{
\item{values}{Values to be marked as outliers}

\item{iqrs}{The number of additional IQRS to extend the range by. Can be
fractional.}
}
\value{
A vector of booleans
}
\description{
An outlier here is defined as any value that is outside of the (25\%, 75\%)
quartile range, plus (or minus) iqrs*IQR. If iqrs = 0, values outside of
(25\%, 75\%) will be marked as outliers. If iqrs = 1, values outside of ((iqrs
\itemize{
\item IQR) - 25\%), (iqrs * IQR) + 75\%) will be marked as outliers
}
}
\examples{
set.seed(1)
values <- rnorm(100, sd = 10)

mark_outliers(values, iqrs = 0)
mark_outliers(values, iqrs = 1.5)
}
