% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{nls_fit}
\alias{nls_fit}
\title{Fit dose response data with best-guess starting parameters}
\usage{
nls_fit(data, cleaned_parameters)
}
\arguments{
\item{data}{data.frame. Has two columns:
\itemize{
\item dose: numeric, contains untransformed doses
\item response: numeric, contains responses as inhibition
}}

\item{cleaned_parameters}{list containing:
\itemize{
\item params: a named vector of fit coefficients
\item max_lower, min lower, and max_upper: additional constrains to provide nls
if slope is shallow
}}
}
\value{
nls fit
}
\description{
Fit dose response data with best-guess starting parameters
}
