% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Check data integrity and format for analysis}
\usage{
preprocess_data(data, is_viability)
}
\arguments{
\item{data}{data.frame containing columns Conc1, Conc2, and Response, all of
which should be numeric}

\item{is_viability}{boolean. Is the Response column measuring viability? If
not, assumes measuring inhibition, and will be converted to viability.}
}
\description{
This function checks for the presence of columns as well as if they are of
appropriat type, and converts Response to viability if is_viability is FALSE.
}
