% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrease.R
\name{decrease}
\alias{decrease}
\title{Impute missing values from dose-response experiments}
\usage{
decrease(path, is_viability, use_fitted_single_agent_values)
}
\arguments{
\item{path}{character path to .csv file containing columns Conc1, Conc2, and
Response}

\item{is_viability}{boolean. Is the Response column viability? If FALSE,
converted to viability.}

\item{use_fitted_single_agent_values}{boolean. Should the fitted single-agent values be
used? If FALSE, use raw values.}
}
\description{
Impute missing values from dose-response experiments
}
